requirements.txt
flask
openai
requests
websockets
python-dotenv
twilio
mcp
pydantic
anthropic
pinecone-client
tiktoken
-e ./servicenow-mcp
